towards filter relocation
