+++ worker/walrus/src/filters/table/publication.rs
@@ -1,0 +1,34 @@
+use cached::proc_macro::cached;
+use cached::TimedSizedCache;
+use diesel::*;
+
+pub mod sql {
+    use diesel::sql_types::*;
+    use diesel::*;
+
+    sql_function! {
+        #[sql_name = "realtime.is_in_publication"]
+        fn is_in_publication(schema_name: Text, table_name: Text, publication_name: Text) -> Bool;
+    }
+}
+
+#[cached(
+    type = "TimedSizedCache<String, Result<bool, String>>",
+    create = "{ TimedSizedCache::with_size_and_lifespan(250, 1)}",
+    convert = r#"{ format!("{}.{}-{}", schema_name, table_name, publication_name) }"#,
+    sync_writes = true
+)]
+pub fn is_in_publication(
+    schema_name: &str,
+    table_name: &str,
+    publication_name: &str,
+    conn: &mut PgConnection,
+) -> Result<bool, String> {
+    select(sql::is_in_publication(
+        schema_name,
+        table_name,
+        publication_name,
+    ))
+    .first(conn)
+    .map_err(|x| format!("{}", x))
+}
